SUS-3104: Theme Designer work around legacy hack
